feat: python oop exercises
