--- m4-ciencia-da-computacao/bloco-34-padroes-de-projeto/01-p.o.o-em-python/learning-exercises/person.py
@@ -1,5 +1,0 @@
-class Person:
-    def __init__(self, name, balance):
-        self.name = name
-        self.balance = balance
-        self.refrigerator = None
